fixed64: codemod pass 9 -- cascade scalar narrowings + revert a mis-wrap
17 more Fixed->scalar casts (astar LONG coords, etc.). One codemod
mis-wrap (a Draw() parameter declaration got wrapped as (int)(int alpha),
which cascaded to its call sites) -- reverted; the flag-off build gate
caught it before it could ship. A diff-wide scan confirms it was the only
declaration mis-wrap. Flag-off build green; tests green.

+++ src/zandronum/src/astar.cpp
@@ -797,7 +797,7 @@ static ASTARNODE_t *astar_GetNodeFromPoint( POS_t Point )
 //
 static LONG astar_GetCostToGoalEstimate( ASTARPATH_t *pPath, ASTARNODE_t *pNode )
 {
-	return ( P_AproxDistance( pNode->Position.x - pPath->pGoalNode->Position.x, pNode->Position.y - pPath->pGoalNode->Position.y ) / FRACUNIT );
+	return (LONG)(( P_AproxDistance( pNode->Position.x - pPath->pGoalNode->Position.x, pNode->Position.y - pPath->pGoalNode->Position.y ) / FRACUNIT ));
 /*
 	POS_t	PosGoal;
 	POS_t	PosNode;
